For Sakai 2.0.1 and CTools 2.2
Fix to SAK-841: Email Archive Tool's content should be imported when using Site Archive Tool.

The archived email message is marked from "postmaster". Since "postmaster" is not a user in the target site, no email message is imported.

The fix ignores the user check when the message is marked from "postmaster".