Related to SAK-1417
Making portal more accessible.  Changed the IFRAMEs to have "title" attribute.  The title
attribute of the title frame is the tool title.  The title attribute of the main frame of 
the tool is toolTitle+" Content". 

For example, the Announcements tool titlebar frame will have the title "Announcements", and the
Announcements tool main tool frame will have the title "Announcements Content".