SAK-1838
Improved the user cache.  It is OFF by default - enable it in the sakai configuration.
Added control over the cache cleaner frequency, run it more often to reclaim memory faster (at increased processing cost).
Added invalidation of the user records in the cache when that record is modified (across the cluster).

A new memory cache is now available for a non-refresh, timed and event driven expiration.