This adds the feature requested in SAK-3465
The following URLs are now supported in Charon:

http://localhost:8080/portal/page/sakai.resources?sakai.site=~csev
http://localhost:8080/portal/tool/sakai.resources?sakai.site=~csev
http://localhost:8080/portal/title/sakai.resources?sakai.site=~csev

In this form, the specified site is scanned and the first placement of the tool within 
the site is the one chosen.  This works to replace either the page id or the tool id 
for these three forms (page, tool, title).  When the page form is used, the definition 
is to select the page that has the specified tool placed on that page.  If there are 
more than one tool is on the page, all the tools on the page are rendered.  Unless 
the site allows anonymous visitors, the URL pattern demands the user to be 
logged in prior to looking up the site.