Fixes SAK-3482, SAK-3483
This fixes a problem when connecting from the Windows XP WebDAV client when the 
Sakai is connected behind an Apache through mod_jk.  The Windows XP client just 
hangs.   This modification catches the problem and returns a more reasonable 
return message so as not to hang the Windows client.

This should go into the maintenance branch for 2.1.1