Implement Rutgers changes to Access servlet including aliases of the form
  http://localhost:8080/access/~hedrick to find use directories
  http://localhost:8080/access/alias where alias is the site's alias (currently mail alias)

Also implements POSTing of resources via Access Servlet (so this is a form of upload)

Also implemnts the serving of index.html if the URL is a directory.

If the URL is a directory and there is no index.html generate a compliant list of files similar to Apache.

Adds support for generating generating a file picker in Sferyx format as well.

As complex as these changes seem, they are all "upwards" compatible (i.e. adding new functionality).

This has been tested and is in production at several sites, but we should pay some attention to testing this
new functionality during our next QA cycle (2.1.2).

Thanks a lot to Chuck Hedrick's hard work and others as well for this excellent set of features.

For more detail see SAK-3442