svn merge -r 5775:5777 https://source.sakaiproject.org/svn/trunk/sakai/ . U    legacy/component/src/java/org/sakaiproject/component/legacy/content/DbContentService.java U    legacy/dav/src/java/org/sakaiproject/dav/DavServlet.java svn log -r 5776:5777 https://source.sakaiproject.org/svn/trunk/sakai/ . ------------------------------------------------------------------------ r5776 | csev@umich.edu | 2006-02-05 00:55:32 -0500 (Sun, 05 Feb 2006) | 5 lines
Commit Seth's changes so as to return a 201 when MKCOL is successful rather than the 200 error code.

SAK-3640


------------------------------------------------------------------------
r5777 | csev@umich.edu | 2006-02-05 01:03:02 -0500 (Sun, 05 Feb 2006) | 27 lines

Fix the 2.1.x WebDav problems on the Macintosh.

The problem only appears when blobs are stored in the file system and 
using the Mac 10.4 client.

The underlying problem is in ContenxtHosting where files of zero length 
were treated differently in the file system storage versus in the DB storage.

In DB storage (for a long time now) - it is non-fatal to try to store 
a zero length content body - the row is never added but there is no 
error condition.  In DB storage when reading when there was no body, 
it was non-fatal and either a null stream was returned or an empty byte array.

Because of some new error checking code added to the storing of blobs in 
the file system, ContentHosting reports ServerOverload in the case where there is 
an attempt to store a zero length blob.  

This modification changes the storage of a zero length blob in the file 
system to be non-fatal.  No file is created (paralleling how the DB works).
Code was added to the blob retrieval code to check for all of the cases 
where the blob did not exist.  However if the resource's metadata 
indicates that a file has length > 0 *and* there is not data to read
(i.e. the file is gone) it is logged.

SAK-3640.


------------------------------------------------------------------------