svn merge -r 5139:5140 https://source.sakaiproject.org/svn/trunk/sakai/ . U    docs/architecture/sakai_charon.doc U    portal/charon/src/java/org/sakaiproject/portal/charon/CharonPortal.java svn log -r 5140 https://source.sakaiproject.org/svn/trunk/sakai/ . ------------------------------------------------------------------------ r5140 | csev@umich.edu | 2006-01-06 11:07:34 -0500 (Fri, 06 Jan 2006) | 27 lines
This removes the logout option from gallery mode.  Since gallery mode is generally 
included in some other framework such as a portal, there is no clear need for the 
gallery mode to support logout.  

This adds a new option which determines whether or not gallery mode is routed through login.  
The prior behavior is that when a browser navigates to /gallery, they are shown the gateway 
site because they are not logged in.  This makes little sense because the typical use case 
is to send the user to gallery so that they can actually use Sakai as a logged in user.  If 
Sakai is not using single sign-on and the portal is not using session-based launch and the user 
is sent to /gallery they are asked to log in.  This behavior is the same as if the user 
navigates to a site or tool url.

The prior behavior (sending the user to a gallery version of the gateway site) can be caused 
by adding a force.login parameter as follows /portal/gallery?force.login=no

These modifications should not change any behavior if WEBISO is being used for Sakai, or if 
session based launching is being used.

This fixes the following bugs:

SAK-3119 Gallery mode should have optional logout
SAK-3118 Logout from Gallery mode returns user to site mode
SAK-3120 IFrame integration of Sakai is problematic with current url-based site navigation

This should be moved to the maintenance branch and included in 2.1.1


------------------------------------------------------------------------