svn merge -r 5260:5261 https://source.sakaiproject.org/svn/trunk/sakai/ . U    legacy/dav/src/java/org/sakaiproject/dav/DavServlet.java svn log -r 5260:5261 https://source.sakaiproject.org/svn/trunk/sakai/ . ------------------------------------------------------------------------ r5261 | csev@umich.edu | 2006-01-11 16:48:09 -0500 (Wed, 11 Jan 2006) | 11 lines
Fixes SAK-3482, SAK-3483

This fixes a problem when connecting from the Windows XP WebDAV client when the 
Sakai is connected behind an Apache through mod_jk.  The Windows XP client just 
hangs.   This modification catches the problem and returns a more reasonable 
return message so as not to hang the Windows client.

This should go into the maintenance branch for 2.1.1



------------------------------------------------------------------------