svn merge -r 5349:5350 https://source.sakaiproject.org/svn/trunk/sakai/ . U    legacy/dav/src/java/org/sakaiproject/dav/DavServlet.java svn log -r 5350 https://source.sakaiproject.org/svn/trunk/sakai/ . ------------------------------------------------------------------------ r5350 | csev@umich.edu | 2006-01-16 01:57:32 -0500 (Mon, 16 Jan 2006) | 9 lines
Change logging level for some normal use cases.  The use case is that the
Windows XP DAV client checks for a file to exist before creating it.

Prior to this mod this generated a WARN in the log.  This is a completely normal 
occurance so the proper log level is debug.

This is additional cleanup for SAK-3303.


------------------------------------------------------------------------