svn merge -r 5779:5780 https://source.sakaiproject.org/svn/trunk/sakai/ . U    legacy/dav/src/java/org/sakaiproject/dav/DavServlet.java svn log -r 5780 https://source.sakaiproject.org/svn/trunk/sakai/ . ------------------------------------------------------------------------ r5780 | hedrick@rutgers.edu | 2006-02-05 16:34:02 -0500 (Sun, 05 Feb 2006) | 18 lines
May fix SAK-3483

The original fix broke propfind, causing it to ignore the user's
request for what data to return. The problem this was fixing is a hang
hang from XP in a mod_jk environment. I've been unable to duplicate
the problem, so I can only guess at a better fix. The old code passed
the XML parser an input stream on the httpservletrequest. Instead I'm
reading the data into a buffer and passing the parser a stream on the
buffer. Because the code for reading into a buffer works for PUT, I
conjecture that this approach will work.

If this fails, I'd like to try to duplicate the hang. I've tried
creating a network place on XP, using both normal and SSL connections
through mod_jk. This is with OS X as the server. I haven't seen the
hang. If you could give me enough information to duplicate the hang,
I'd appreciate it.


------------------------------------------------------------------------