svn merge -r 5348:5349 https://source.sakaiproject.org/svn/trunk/sakai/ . U    legacy/adaptor/src/java/org/sakaiproject/component/adaptor/UsageSessionServiceAdaptor.java svn log -r 5349 https://source.sakaiproject.org/svn/trunk/sakai/ . ------------------------------------------------------------------------ r5349 | csev@umich.edu | 2006-01-16 01:46:24 -0500 (Mon, 16 Jan 2006) | 9 lines
This changes the startSession code to remove the WARN message when 
the startSession is called when a UsageSession already exists.  The
existing session is returned if the UserId of the current session matches
the UserId of the "new" session.  If these two do not match, a brand new 
session is created and a warning message is emitted.

Fixes SAK-3303.


------------------------------------------------------------------------