CM-37 -- Finished initial implementation of the CourseManagementGroupProvider, including section, course offering, and course set role resolvers.  Cross-listing still needs more work, and everything needs more testing.
In order to complete the group provider (which is, in a real sense, the first client for CM), I had to make some changes to the API:
o removed getSectionRole(String sectionEid, String userEid) since it's not needed
o added findSectionRoles, findCourseOfferingRoles, and findCourseSetRoles methods to satisfy the group provider.  These methods should be renamed, once we come up with more suitable ones.