svn merge -r 13524:14333 https://source.sakaiproject.org/svn/assignment/trunk U    assignment-tool/tool/src/webapp/vm/assignment/chef_assignments_list_assignments.vm U    assignment-impl/impl/src/java/org/sakaiproject/assignment/impl/BaseAssignmentService.java /Develop/maint_2-2-x/assignment_2-2-x$ svn log -r 13524:14333 https://source.sakaiproject.org/svn/assignment/trunk ------------------------------------------------------------------------ r14332 | zqian@umich.edu | 2006-09-11 13:51:48 -0400 (Mon, 11 Sep 2006) | 1 line
Fix to SAK-6259: group aware assignments get deleted if the group is deleted and leave an error message that cannot be taken out of the assignment page.
------------------------------------------------------------------------
r14333 | zqian@umich.edu | 2006-09-11 13:51:51 -0400 (Mon, 11 Sep 2006) | 1 line

Fix to SAK-6259: group aware assignments get deleted if the group is deleted and leave an error message that cannot be taken out of the assignment page.
------------------------------------------------------------------------