SAK-2440
Allows one to parameterize the Sakai Services and Roles that are Merged into
a site via the Spring components file.

This fix includes a new project, impl2, which rips apart a few things from
the original. The original is left as is, because support for legacy systems
such as WorkTools and CourseTools Classic is removed in the impl2.

This also includes a change to the components.xml in archive-pack.  The original
bean is left in the file, commented out, in case it needs to be used in the
future by a school upgrading from one of the legacy systems.