http://bugs.sakaiproject.org/jira/browse/SAK-6301 http://bugs.sakaiproject.org/jira/browse/SAK-5937
Added checking of invalidEmailInIdAccountString to the audience tool
Added multiple domain capabilities to invalidEmailInIdAccountString (it is a comma separated list where spaces are trimmed, e.g. "indiana.edu,iupui.edu, iu.edu")
This multi-domain capability is in both the osp audience tool and in the SiteAction where emails can be added.