svn merge -r 14941:14942 https://source.sakaiproject.org/svn/site-manage/trunk U    site-manage-tool/tool/src/java/org/sakaiproject/site/tool/SiteAction.java /Develop/maint_2-2-x/site-manage_2-2-x$ svn log -r 14941:14942 https://source.sakaiproject.org/svn/site-manage/trunk ------------------------------------------------------------------------ r14942 | andersjb@iupui.edu | 2006-09-19 14:44:58 -0400 (Tue, 19 Sep 2006) | 6 lines
http://bugs.sakaiproject.org/jira/browse/SAK-6301
http://bugs.sakaiproject.org/jira/browse/SAK-5937

Added checking of invalidEmailInIdAccountString to the audience tool
Added multiple domain capabilities to invalidEmailInIdAccountString (it is a comma separated list where spaces are trimmed, e.g. "indiana.edu,iupui.edu, iu.edu")
This multi-domain capability is in both the osp audience tool and in the SiteAction where emails can be added.
------------------------------------------------------------------------