http://bugs.sakaiproject.org/jira/browse/SAK-6229
when submitting a wizard page or a matrix page from the confirmation screen, the user is returned to the proper screen.
The hierarchical page submission was being directed to the sequential page
cleaned up some views

To test:
have a wizard (both seq and hier) and a matrix published. go in as a user and fill in a page in the three respective areas and click the submit to evaluate.  then on the confirmation page click submit.  It shouldn't error in this last step.