svn merge -c 15519 https://source.sakaiproject.org/svn/assignment/trunk U    assignment-tool/tool/src/java/org/sakaiproject/assignment/tool/AssignmentAction.java /Develop/maint_2-2-x/assignment_2-2-x$ svn log -r 15519 https://source.sakaiproject.org/svn/assignment/trunk ------------------------------------------------------------------------ r15519 | zqian@umich.edu | 2006-10-03 11:35:52 -0400 (Tue, 03 Oct 2006) | 14 lines
Fix to SAK-6607:'Permissions' not working in Assignments if the 'Student view' tab is clicked

Fixed the problem of 

"While the above bug is present, clicking on the 'Grade report' tab and then clicking on Permissions returns the following message 

' Assignment List 

There are currently no assignments at this location. ' "

However, the "Permission" link not working is not limited in Assignment tool alone. It appears in Announcement tool, Schedule tool, etc. 
It happens after user clicks on the Reset button of the tool and then clicks on the Permission button. 

By checking the HTML source code for the toolbar links, the panel attribute of the link has null value, which causes the above problem. 
------------------------------------------------------------------------