svn merge -r 15837:16924 https://source.sakaiproject.org/svn/osp/trunk U    common/component/src/webapp/WEB-INF/components.xml U    common/api-impl/src/java/org/theospi/portfolio/style/impl/StyleManagerImpl.java U    common/api-impl/src/java/org/theospi/portfolio/worksite/mgt/impl/WorksiteListGenerator.java U    common/api/src/java/org/theospi/portfolio/style/model/Style.java U    common/api/src/java/org/theospi/portfolio/style/model/Style.hbm.xml U    common/tool/src/bundle/org/theospi/portfolio/common/bundle/Messages.properties U    common/tool/src/webapp/audience/steps.jspf U    wizard/tool/src/webapp/wizard/runWizardGuidance.jsp U    matrix/tool/src/webapp/WEB-INF/matrix-web-config.xml U    presentation/api-impl/src/java/org/theospi/portfolio/presentation/export/PresentationExport.java U    presentation/tool/src/bundle/org/theospi/portfolio/presentation/bundle/Messages.properties U    presentation/tool/src/bundle/org/theospi/portfolio/presentation/bundle/Messages_sv.properties U    presentation/tool/src/webapp/WEB-INF/jsp/presentation/listPresentation.jsp /Develop/2-3-x/osp$ svn log -r 15837:16924 https://source.sakaiproject.org/svn/osp/trunk ------------------------------------------------------------------------ r16905 | chmaurer@iupui.edu | 2006-10-09 08:27:37 -0500 (Mon, 09 Oct 2006) | 2 lines
http://bugs.sakaiproject.org/jira/browse/SAK-6021
A hash is now stored with each style item and will be consulted when importing a new style.  If the same hash is found for a global style, or in the same site, the style will now be imported.
------------------------------------------------------------------------
r16908 | chmaurer@iupui.edu | 2006-10-09 08:49:37 -0500 (Mon, 09 Oct 2006) | 2 lines

http://bugs.sakaiproject.org/jira/browse/SAK-6771
The site list synoptic will now take a list of site types (or none for all types) to return.
------------------------------------------------------------------------
r16917 | andersjb@iupui.edu | 2006-10-09 14:15:17 -0500 (Mon, 09 Oct 2006) | 3 lines

http://bugs.sakaiproject.org/jira/browse/SAK-6622

Added some error checking with better output.
------------------------------------------------------------------------
r16918 | andersjb@iupui.edu | 2006-10-09 14:19:22 -0500 (Mon, 09 Oct 2006) | 6 lines

http://bugs.sakaiproject.org/jira/browse/SAK-5993

Changed the portfolio text to read Share instead of Publish

To test:
have a portfolio in the portfolio tool. It should say Share instead of Publish.  When clicking share and when going to the third step in the revise, all the text should say share and not publish
------------------------------------------------------------------------
r16924 | andersjb@iupui.edu | 2006-10-09 17:14:42 -0500 (Mon, 09 Oct 2006) | 3 lines

http://bugs.sakaiproject.org/jira/browse/SAK-6781

Changed the method of getting attachments so that they always bypass security
------------------------------------------------------------------------