http://bugs.sakaiproject.org/jira/browse/OPC-50
when the forums in the area aren't loaded and when the topics in the forums aren't loaded then the saving code ignores the reindexing check and reindexing.

This change was made because in some cases the areas and forums were saved without having their forums and topics respectively loaded.  When this happened the reindexing code tried to access the set which caused a LazyInitException because the session that originally loaded the area or forum was closed.