Begin work on the dense portal (aka PDA portal, aka the portlet portal).  The basic idea is to flatten all navigation up to a single line - no tabs or buttons.  Just simple flat lists.   Implicit reset throughout.
For now this is only in the Velocity branch of Charon.  There is more work to do including skinning, login/logout, 
and refactor some of Ian's code to use some more generalized and reusable routines that follow a more MVC pattern
since we are now in Velocity.

If you want to play with it - get Charon Velocity Branch checked out and get Maven 2 set up - 
kind of tough to look at for now.

SAK-8180