Patch supplied by John Fawcett ------------------------------ Ian,
This is the first of two patches for the problem with the search
tool.  The same problem manifests itself in several guises.  This
particular patch eliminates the problem of being unable to perform
several "edit" actions in Content Hosting as soon as a virtual 
resource has been introduced.  The symptoms were that java spins
at high-ish CPU and takes hundreds of minutes to complete the 
request.  The work being done in my test case was enumerating the
millions of files in my home directory and summing their sizes 
(mod 231).

john