Add features to Sakai's JSR-168 support as documented in
/reference/docs/architecture/sakai-168-portlets.doc

The primary feature is to add more control to the portlet registration process 
including the ability to add a tool registration file for a portlet without modifying the
portlet war file as well as a property to control the auto-regitration behavior.

Requires modifications to tool/ActiveTool up to revision 21132.

SAK-8022
SAK-8349