SAK-8251
Fix for issues with HTML and Text indexing,

This contains a custom HTML parser that uses regex rather than XML and so is far more 
robust that Tidy+SAX.

At the moment it has not been heavilly tested although there is a unit test that throws some reasonably complex
and extreemly broken pages at it.

I have also fixed the issues with binary files that are now filtered on mime type.