SAK-8413
Better control of Logging in POI,

However,

The version of POI we have to use is a snapshot, and it has System.err.println in the code, 
There is no release of this version at the moment, so we are going to have to live with this problem at the moment.
I need poi-snapshot-3.0-alpha3 for things to work better.