Fix bug in JSR-168 where portlets are rendered twice for each request.
First, remove the call to doGet() in the doPost() method for the Site and PDA Handlers.

Second, after the pre-process step in the Portal, check to see if pre-process did a redirect
and end the processing on this request if we were redirected.  Of course the redirect was back to us
as a GET so we will process things properly anyways.