SAK-8988 Since the merging operation has changed, we are not finding that the lucene segment is no longer append only. The individual cfs files within the segment are append only but lucene reorganizes those segments hence all of the checksum failures that we were experiancing. The index distribution mechanism has been modified to unpack into a spare temporary directory, move the contents into the existing segment directory and then remove files that were not in the segment. However... this may not be the best way of doing things as the live and running index may not like having segments removed from under it. We will have to see what happens after some more testing.
I have also broken the segment reload in a cluster as the Sakai Events are not boradcasts we have to poll the 
index status and reload, if there is more than one thread on a node this can cause reloads when the index is changing.
We may need to go to 1 thread per node and make that thread the reload and index thread so that we dont end up with the segment 
in a strange state. This could simplify the segment update mechanism as we could delay deletes untill the segment has been reloaded.

I will be running continuous tests today to make certain that this all works, but its a very slow process getting content into content 
hosting.