SAK-9335
Add an experimental feature to the PDA portal that eliminates iframes for some of the Sakai 
tools using a Buffered Servlet response and parsing the results.  This will need some
experimentation so there is a property that gives some flexibility.

# Indicate the tool Ids that work with the PDA inline feature
# Add the String :debug: to indicate we want to see debug output in the  log
# Do not include debug when in production
# portal.pda.experimental.iframesuppress=debug:sakai.createuser:sakai.motd:sakai.iframe:sakai.ifame.service:sakai.iframe.myworkspace:sakai.iframe.site:sakai.profile:sakai.membership:sakai.resources:sakai.dropbox:sakai.announcements:sakai.sitesetup:sakai.synoptic.discussion:sakai.announcements:sakai.synoptic.announcement:sakai.mailbox:sakai.news:sakai.messages:sakai.synpotic.chat

# Buffer output for all tools - Only use this for testing because some will break
#portal.pda.experimental.iframesuppress=all

# Problems with non-iframe that need testing
# sakai.summary.calendar "< and > seem to fail but right click works - crazy"
# sakai.schedule "Next Day seems to Fail but right click works - crazy"
# sakai.chat (fails badly - too many frames)
# sakai.discussion (fails badly - too many frames)
# sakai.siteinfo (page order fails)
# sakai.rwiki (preview fails)
# blogger
# sakai.poll

Some may suggest merging this into the branch for 2.4 because it only touches an experimental and 
effectively non-tested area of code.  I would say that this would be nice to have in 2.4 and not
harm any part of 2.4.