svn merge -c 28253 https://source.sakaiproject.org/svn/citations/trunk U    citations-tool/tool/src/java/org/sakaiproject/citation/tool/CitationHelperAction.java svn log -r 28253 https://source.sakaiproject.org/svn/citations/trunk ------------------------------------------------------------------------ r28253 | gbhatnag@umich.edu | 2007-04-02 18:36:38 -0400 (Mon, 02 Apr 2007) | 1 line
SAK-9183: Added checks to make sure that the Reference we're getting from EntityManager is not null.  This may be related to a Resources-level issue having to do with the Administration Workspace, but at least now NullPointer's are not being thrown by Citations code and there is a backup in case the Reference is null.
------------------------------------------------------------------------