Initial commit of the WURFL Java APIs.
This has been pulled into the Sakai source tree because
the WURFL code has not been maintained for over 4 years 
and there are flaws in the code.

The biggest problem to solve is dealing with User Agent
values that simply increment a version number.

The "Loose" matching should do this - but instead does
a really poor successive match that threw away
bits of the user agent - to the point where it effectively
picks a random device - I will change this to eliminate
any numbers and then do a match of the agents ignoring
versions.

Also this can clean up loging patterns.