SAK-9799
Make is so that if there is only one tool permitted to be viewed that 
the list of tools is not shown.  This also suppresses presense for the time being.
Presense could be moved into a DIV (like PDA portal) if a UI could be worked out.

Since this is non-upwards compatible UI behavior it is an option for a while.
This will let folks experiment and determine if it is sutiable to always 
do this - or retain this as an option - my suggestion is to make this the 
standard behavior as long as presense can besuitably solved.

# Indicate that where there is only one page viewable on the site - 
# supporess the left column including the tool list and presense
portal.experimental.maximizesinglepage=true