SAK-9889
This patch is much bigger than the previous one, as it needed the URL processing to be sorted out.

I have tested as far as I can, and all appears to be Ok, but it may need more testing.

I dont have FF2 installed which should react to the opensearch urls.