OSP is now converted,
There are a significant number of excludes as the old code was carrying arroung cglib and aop libraries as dependencies.
OSP team should investigate the scope of these dependencies at their source (usually in the tool-lib jars) and see if 
they can be removed as transitives and hence excludes.

There were also a significant number of missing jars in the packs potentially caused by using libs with JSF dependencies in
components. Again these should be investigated and removed at source. Ieally there should be no excludes in the build.

The pom layout is hierachical and singnificant savings could be made in the number of dependencies by using this hierarcy. 
This has not been done with this commit. However you *should not* use dependencyManagement at more than one layer as there
are bugs in maven 2 that only allow a single layer of dependency management on pack.