SAK-10306 SAK-10810
Changed reg exp pattern for detecting urls in Formatted text such that it no longer dies when the input contains things such as 'target="_blank"' please note that as it's been for quite some time if you do not call processFormattedText with checkForEvilTags set to false that it will remove all such things from the input and inforce all links to open with target="blank" anyway. If this is not the desired functionality then we should have a discussion.