SAK-9674
Fixed the UTF-8 problem in the title, but also noticed the logic was wierd, we were html encoding where we should not have been. 
All html encoding has been moved into the view layer, and the use of the highlighter has been fixed so that it correctly highlights UTF8 fragments.