SAK-10366
This is the merge of the SAK-10366 branch.
The JCR Code is isolates and not part of the build at the moment,
content-impl-jcr/impl needs to be activated in the build once the JCR API is in place.

To use you will need to deactivate the standard content-impl-pact and active content-impl-jcr-pack 
which contains both the DB content service and the JCR content service.

You will also need to deploy a JCR implementation into Sakai to provide the JCRService