SAK-10380 http://jira.sakaiproject.org/jira/browse/SAK-10380
Merged,
BUT

Not happy with the use of hibernate and the query mechanism used to load entity properties. 
I see performance problems ahead with this if the Hibernate DAO usage becomes high.

I would like to see some high load concurrency checks to make 100% certain that this does
not cause load problems in the DB and/or deadlocking.