SAK-11544
Add the getCount() and getPagedMessages() methods to the message API and 
add simple versions of these methods to the Base class. 

Currently no tools use these methods and for efficiency they should 
be overridden in the real implemntations. If tools 
do start to use these methods without their services overriding the 
methods they will work as well as the rest of the implementations 
in the Base class. 

The first consumer of these will likely be the Mail Archive tool and service. 

This is not a complete solution - it is getting the APIs right so that 
the implementations can be built in Mail Archive to do this more 
efficiently.