SAK-11245 http://jira.sakaiproject.org/jira/browse/SAK-11245
Soak test now runs for 1 hour witn 1000 journals 4 nodes, and no problems, 
cleaning the segments is a bit complex to avoid earlier problems of deleting files when 
existing segments where open.

I now need to address the list of active segments on the local machine. Since the merger and optimizer
are operating within a write locked space this can be a write to a file on the local filesystem containing a list of segment 
numbers.