SAK-11245 http://jira.sakaiproject.org/jira/browse/SAK-11245
Fixed slightly obscure memory leak, If you add a single shot timer task to a Timer, you have to explicitly cancel it
as part of the run method to ensure that the Timer Queue releases it.

Index, Merger, Optimzer and Searcher now runs for 1 hour with no error on 5 threads.