svn merge -r 37679:37680 https://source.sakaiproject.org/svn/msgcntr/trunk U    messageforums-component-impl/src/java/org/sakaiproject/component/app/messageforums/ui/PrivateMessageManagerImpl.java U    messageforums-app/src/java/org/sakaiproject/tool/messageforums/PrivateMessagesTool.java U    messageforums-app/src/java/org/sakaiproject/tool/messageforums/ui/MessageForumSynopticBean.java U    messageforums-app/src/webapp/jsp/privateMsg/pvtMsg.jsp in-143-196:~/java/2-5/sakai_2-5-x/msgcntr mmmay$ svn log -r 37679:37680 https://source.sakaiproject.org/svn/msgcntr/trunk ------------------------------------------------------------------------ r37680 | hu2@iupui.edu | 2007-11-01 09:03:59 -0400 (Thu, 01 Nov 2007) | 11 lines
SAK-11130
http://jira.sakaiproject.org/jira/browse/SAK-11130
With a localized Sakai, if pvt_deleted, pvt_received and pvt_sent are not set to "Deleted", "Received" and "Sent" (default values). The folders that are not set to default values do not work.

Without looking at the code I think those default folders names have a internal name which is the same as their default label. When the label is localized, the code is unable to match the localized label to the internal name.

Quick fix: do not localize pvt_deleted, pvt_received and pvt_sent. I've done it and it works.

Better but not-so-quick fix: match those default folders localized labels and internal names in the code. I won't complain if someone else does it. :-)

it support english and spanish now.
------------------------------------------------------------------------