SAK-12134
This is a working version of the Lifecycle component loader that loads the component manager based ont eh container lifecycle rather than as a result of the webapp.
It injects the a component manager mbean into the MBeanServer and this is then used by the static cover to recover the Component Manager.

The static cover loads from the MBean and and there is a component manager proxy bean that enables the elimination of the static cover for the component manager
This bean can be created eg new instance and it will associated itself with the component manager on construction.