svn merge -c39208 https://source.sakaiproject.org/svn/content/trunk/ . U    content-tool/tool/src/java/org/sakaiproject/content/tool/ResourcesHelperAction.java
svn log -r39208 https://source.sakaiproject.org/svn/content/trunk/
------------------------------------------------------------------------
r39208 | jimeng@umich.edu | 2007-12-13 08:08:24 -0800 (Thu, 13 Dec 2007) | 3 lines

SAK-10042
File-count was being set to zero when quota was exceeded.  It needs to be >1 to show the input form for uploading a file, and the input form needs to be there for the "add another" link to work.