SAK-11544
Commit of some test code - please don't look at this - it is a bit of refactor in Mail Action
and a bad hack to fill up a mail archive by pressing the options button.  I just wanted to put
this in my happy little branch so I did not have to keep carrying this code around manually.