SAK-12402
Futher cleanup - introduce two new properties:

sakai:portlet-pre-render - indicates that the portal is to pre-render portlet
content before the view starts - this allows the portal to check to see if the 
portlet has requested to be switch to/from maximize mode on 
any particular request.

sakai:prefer-maximize - This indicates that the tool always wants to 
be maximized.  This makes most sense for non-JSR-168 tools that want 
to have a lot of horizontal and vertical space and want portal navigation
to be as minimal as possible while the tool is active.

This branch is getting pretty mature now.  I need to write documentation
about these new features next.  I do need to do a line by line code review
of all my changes since the start of the branch as well before I am happy.
I also am writing a test plan for these features.