SAK-12402
Make it so that when you go from reduced navigation mode back to normal 
navigation mode you go to the first tool instead of the previously selected
tool - this way the navigation is more natural. You don't end up seeing the 
tool you just maximized sitting in an iframe - you likely see the Home tool
or whatever the top tool is instead.