SAK-12511
I think I found the problem that was creating a gazillion timer threads and was
able to successfully migrate a fairly large number files/sites and open them in
the hot swapped impl.  Just need to fix the loop to check the status and do a 
bit more cleanup before running on the test server.

Also made a SakaiEmulatedRequest class that can be extended, based off the 
SakaiTestCase from the TestRunner to fake out the CHS.