SAK-12640
Added a specific Exception LockTimeoutException that extends the JournalIndexExcpetion and then bubbles up through the 
Optimization Transaction Manager and is caught explicity by the JournalOptimizationOperation so that it doesnt generate
a big traceback in the logs.

This does not change the logic of the locking of otimization records, it just changes the lock timeout handling and reporting.