SAK-12642
Fixed,
If the local index was zapped losing local state and the database remained, when the server ID was re-used there would still be a 
record in the database, hence when no local state is found the database state will be reset.

If server ID's are shared between nodes, havoc will result.

Longer term, perhapse we should be storing the local journal pointer in the local segments file to make certain that the merged mainsegment 
only contains what we expect it to. I see problems ahead if a dirty local index is used. However, its really easy to rebuild, just remove local-segments and restart the node.