SAK-12706
Added control over the index merge parameters that should make it possible to change the structure of the index.
Currently set to have 10 segments per journal with a factor of 10, see documentation inside parallelIndexComponents.xml for 
more information.

This bug is not closed as I want to look and see if I can optimize the shared merge/optimize further.