NOJIRA Added portal http testing framework
At the moment, this  spders a known set of seed urls and posts to the login form when presented as admin.
It only starts if you have a sakai instance on http://localhost:8080 and at the moment it doesnt look for errors.

As this develops we will add checks for errors. 

It does load the iframes of tools and will spider into iframes, but inless more ActionHandlers are written it will not 
post any data. It makes an assumption that get urls are Ok to get.