SAK-12387
This is an initial commit of the refactor - this is compatible with trunk - it
could be dropped into trunk.

It includes a number of TODOs that add method signatures that require code
changes elsewhere.  These will be added in a second step where the 
other Storage implementations will be modified to have the new methods.