SAK-11544
Clean up and align with SAK-12837

Remove methods in BaseMailArchiveService that we should inherit 
from BaseMessageService  - after all will just just override 
this stuff in DbMailArchiveService where the real work gets
done.  Makes this a lot easier to distribute and use with trunk
code.  All we need is the SAK-12837 mods that are compatible with
trunk.  This means the patch can be compiled with Sakai as a whole.

Fix some indenting in DbMailArchiveService