NOJIRA
Added Site Content Producer that generates lucene documents for published, public and joinable sites.
Thes documents have a siteid of .auth or .anon depending on if they are public or just joinable. If they are 
neither or the site is not public then the siteid field is .private , this means that these lucene documents
will not be returend in normal queries. 

Is sites are to be searched then the search is +(+siteid:.auth +siteid:.anon) +content:searchterm if the user
is logged in, and +(+siteid:.anon) +content:searchterm if the user is not logged in.

All the other metadata associated with the site entity is available as seperate index keys in the index and can be searched
on if required, using the same index selection mechanism as above.