SAK-11544
Fix the mail archive indexing to not fill up memory with large mailarchive placements.
Instead page through 100 elements at a time, handing them back to the search queue.

This was based on 2-5-x so I had to adjust the POMs back to SNAPSHOT so it can be checked out
and used with the rest of trunk.

I am not entirely sure that mail indexing is seeing the right text from mail messages.  I need
to check this.  What is now fixed is that iindex  large mail archives is done 
responsibly w.r.t. memory.