SAK-13441: The concept of a "preferred URL" has been introduced at the OSID level.  This allows an OSID search component to suggest that a URL it provides is the "best" URL for a result asset.  Note that implementing this feature at the OSID level is entirely optional.
The Citation Tool was updated to use the preferred URL (if available) as the title link.  Normally, an OpenURL would be used for this: the OpenURL is still available as a small link shown below the title link.

The Sirsi Web2 OSID was updated to provide a small "reference" implementation - see PreferredUrlhandler.java for details.