SAK-13834
Terracotta requires custom class loaders to implement the
__tc_getClassLoaderName() method.  TerracottaClassLoader.java extends the
URLClassLoader sakai was already using to both implement this method and
do the correct stuff to get the ClassLoader registered with Terracotta.

ComponentsLoader was modified to look for a system property (could not use
a sakai.property value because of we'd have a circular dependency with another jar file) to know whether we were in a terracotta clustered environment.  If we
are, then it uses the new TerracottaClassLoader, otherwise it does exactly what it was doing before.