SAK-13324
MySession, MyLittleSession and MySessionBindingEvent were made top level
classes, instead of being (non-static) inner classes of SessionComponent.
In order to accomplish this, a couple other changes were required.

All fields that MySession and MyLittleSession used in there previous parent
surrounding class, had to be made member attributes.  These member attributes
had to be populated with constructor parameters.

MySession needed access to the main session collection in SessionComponent
called 'm_sessions'.  Instead of providing this reference directly to a class
(now) outside of SessionComponent, a new mix-in interface, SessionStore was
created, to handle any 'changes' that needed to be made to the m_sessions
data structure.  The MySession constructor now, also requires a SessionStore
object as an argument to its constructor.

In general Session id creation was moved out of MySession, to the
SessionComponent class itself.  This was done, to better define the
responsibilities of SessionComponent.

Given all these changes, the unit tests had to be updated to handle the Mock
objects being created to correpond to the new class hierarchy.